+++ assets/src/components/form/index.js
@@ -1,1 +1,1 @@
-export * from './select';
+export * from './model';
