+++ assets/src/components/form/model/base/default-select-configuration.js
