+++ assets/src/components/form/model/index.js
@@ -1,0 +1,3 @@
+export { ModelSelect, default as ModelEnhancedSelect } from './base/model-select';
+export * from './select';
+export * from './base/build-options';
