+++ assets/src/components/form/model/select/ticket.js
@@ -1,7 +1,7 @@
 /**
  * Internal imports
  */
-import ModelSelect from '../model-select';
+import ModelSelect from '../base/model-select';
 import { ticketModel as model } from '../../../../data/model';
 import { withBaseControl } from '../../../../higher-order-components';
