--- assets/src/components/form/select/index.js
@@ -1,3 +1,0 @@
-export { ModelSelect, default as ModelEnhancedSelect } from './model-select';
-export * from './model-selects';
-export * from './build-options';
