+++ assets/src/components/index.js
@@ -1,6 +1,6 @@
 // query components
 export { QueryLimit } from './query/limit';
 // selection components
-export * from './form/select';
+export * from './form/model';
 // general ui components
 export * from './ui';
